requirements.txt
aiofiles==23.2.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
auth==0.5.3
bcrypt==4.1.2
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
databases==0.8.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
eventlet==0.34.3
exceptiongroup==1.2.0
falcon==3.1.3
fastapi==0.109.0
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.1
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
JPype1==1.5.0
konlpy==0.6.0
lxml==5.1.0
Mako==1.3.0
MarkupSafe==2.1.3
mecab-python==1.0.0
mecab-python3==1.0.8
mongoengine==0.27.0
numpy==1.26.3
packaging==23.2
passlib==1.7.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
py-hanspell==1.1
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==4.6.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.51
starlette==0.35.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
